+++ NEWS.md
@@ -1,6 +1,7 @@
 httpuv 1.5.4.9000
 =================
 
+* Fix SHA1 calculation, and thus WebSocket server handshakes, on big-endian systems.
 
 httpuv 1.5.4
 ============