+++ NEWS.md
@@ -5,6 +5,8 @@ httpuv 1.5.4.9000
 
 * Fixed #195: Responses required `headers` to be a named list. Now it can also be `NULL`, an empty unnamed list, or it can be unset. (#289)
 
+* Allow responses to omit `body` (or set it as `NULL`) to avoid sending a body or setting the `Content-Length` header. This is intended for use with HTTP 204/304 responses. (#288)
+
 httpuv 1.5.4
 ============